rename memEngine to mem_engine
